+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## 2.2.2 (unreleased)
 - [PR #96](https://github.com/otoolep/rqlite/pull/96): Add build time to status output.
+- [PR #101](https://github.com/otoolep/rqlite/pull/101): Fix restore to in-memory databases.
 
 ## 2.2.1 (April 19th 2016)
 - [PR #95](https://github.com/otoolep/rqlite/pull/95): Correctly set HTTP authentication.